pyproject.toml
[tool.poetry]
name = "map"
version = "0.1.0"
description = "A place to store the location of your friends"
authors = ["Weston Gorczyca <weston.gorczyca@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
waitress = "^3.0.0"
python-dotenv = "^1.0.1"
flask-login = "^0.6.3"
requests = "^2.32.3"
flask-sqlalchemy = "^3.1.1"
logging = "^0.4.9.6"
geoalchemy2 = "^0.15.2"
shapely = "^2.0.6"

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
